add the signature of APDUs without any data
Response APDUs of GenKeyPair, StorePubKeyHash and StorePubKey requests don't contain any data, just a return code.

Now, they have the following syntax:
	2B 0x9000 | card's nonce | signature | (return code)

A host sends a 10 bytes long nonce which is used for card's APDU signature. Besides this, the card generates its own 10 bytes long nonce that is also used for the
APDU signature and is sent back to the host so that the signature can be verified.

This should prevent the replay attack.